Move visualizer to Annotation.
