+++ optical/converter/__init__.py
@@ -10,6 +10,8 @@ from typing import Optional, Union
 from .coco import Coco
 from .csv import Csv
 from .yolo import Yolo
+from ..visualizer.visualizer import Visualizer
+from .utils import get_image_dir, ifnone
 
 SUPPORTED_FORMATS = {"coco": Coco, "csv": Csv, "yolo": Yolo}
 
@@ -48,3 +50,12 @@ class Annotation:
             return
 
         self.formatspec.convert(to.lower(), output_dir=output_dir, **kwargs)
+
+    def visualizer(
+        self,
+        images_dir: Optional[Union[str, os.PathLike]] = None,
+        split: Optional[str] = None,
+        img_size: Optional[int] = 512,
+    ):
+        images_dir = ifnone(images_dir, get_image_dir(self.root))
+        return Visualizer(images_dir, self.formatspec.master_df, split, img_size)
